requirements.txt
fastapi
uvicorn
firebase-admin
python-dotenv
pytest
httpx
